Change implement NFSS tables as arrays
First step in preparing for sorting the entries in fd files by weight and width:
rewrite the %NFSS_* tables as arrays, so that we keep the original order of the keys
and may use these to determine sorting order.

We then use "%NFSS_WEIGHT = @NFSS_WEIGHT" to convert the arrays to hashes.