[autoinst] Fix error checking of calls to external programs
When doing something like

```
open my $otfinfo, '-|', qq(otfinfo --info $fontfile);
...
close $otfinfo;
```

the call to `open()` will fail *only* when we couldn't create a child
process (i.e., when `fork()` failed). If the external program run in the
child process encounters an error, it is the call to `close()` that will
fail, so we should test that.

See `perldoc -f open` and the section "Using open() for IPC" in
`perldoc perlipc`.