[autoinst] Improve -mergeshapes
Previously, we only tested the font's family for 'SC' or 'SmallCaps'
when merging separate small caps fonts into the main family;
now we also test the font's name. This was triggered by Underware's
Dolly, where the small caps font has 'family' Dolly and 'name'
Dolly-SmallCaps.

Additionally, we now also remove a dash, if any, between the name or
family and 'SC' or 'SmallCaps'.