[autoinst] Work around bug in font info parsing
Previously, we stripped the 'OT' from the family name 'KievitOT';
but after introducing -mergeshapes, this family name was then
interpreted as the 'italic' fonts from the 'Kiev' family ...

As a workaround, we now leave the OT in place. But maybe it would be
better to improve the handling of -mergeshapes.