[autoinst] Change how 'otftotfm' is called
In commit 2081cc6, we had changed this from system() to IPC:Open3() plus
waitpid(), to allow us to capture stdout and stderr. In turns out,
however, that waitpid() tends to cause the process to hang on Windows.

So now we use open() to a pipe, with a command that includes a redirect
of stderr to stdout, so that reading from the pipe captures both.
According to the interwebs, the "2>&1" syntax works on Windows as well.