[autoinst] Better handling of 'Text' in family name
In font families whose named ended in 'Text' (e.g., Libre Caslon Text),
the 'Text' part would be mistaken for a weight. We had a crude fix for
that; it's been improved.
The earlier fix included stripping of any 'Text' from font family names;
that's unnecessary now, but we still do it to avoid breaking other
packages that rely on the old behaviour.