[autoinst] Overhaul creation of Ornament fonts
Previously, we search the font for ornament glyphs and created
a custom encoding file to contain these, starting from position 1;
but this meant missing ornament glyphs that didn't have 'standard'
names. In the new situation, we just create LY1 encoded fonts (but
specify the NFSS encoding as U) and pass the 'ornm' feature to otftotfm.
This will replace a number of common characters (according to the
standard only lowercase ASCII and bullet, but in some fonts also figures
and maybe uppercase as well) with ornament glyphs. This approach is less
brittle than our previous one, and less work as well :-)