[autoinst] Comment out some math symbol declarations
Some characters turned out to be "repeat offenders"; in many fonts they
dn't really work as math symbols, either because of size, spacing or
because they didn't match with other symbols. For now, we leave these
declarations in the generated style file (albeit commented out), so that
the user may reactivate them if she wants to.