[autoinst] "Reverse" mapping of NFSS codes in .fd
Previously, we installed Regular etc fonts as 'regular' and then
aliased 'm' to 'regular'; this commit ditches the alias and instead
install the Regular fonts as 'm' as well.
Ideally we should install the Regular font as 'm' and then alias
'regular' to 'm' (at least according to Frank Mittelbach). That's
something to think about. For now, let's consider this an experimental
branch.