[autoinst] Test existence before calling Cwd::abs_path
User-specified encodings can be either the *name* of one of our standard
encodings, or the filename (maybe without a filename extension) of a
custom encoding file. We used Cwd::abs_path to turn the filename into an
absolute path, but did that before testing whether the argument actually
referred to an existing file (as opposed to the name of an encoding).
This led to a program crash on Windows. So now we test for existence
before calling Cwd::abs_path.