[autoinst] Fix issue with argument quoting on Windows
A few commits back, we added support for encoding files outside the
current working directory by always quoting the argument to otftotfm's
--encoding option. This does, however, seem to cause problems on Windows
when this argument isn't a full filename but only the name of one of the
included encoding files. So now we only quote this argument when it
actually is a filename.