[autoinst] Fix the fix from 00f21ee
It turned out that Source Serif 4 isn't the only family requiring
this hack; some other (Adobe) families, such as Arno, also need it.
On the other hand, Libertinus Serif Display is the only known
counterexample; so it makes more sense to apply the hack by default
and have Libertinus as an exception.